% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGrowthRate.R
\name{GetGrowthRate}
\alias{GetGrowthRate}
\title{Calculate the growth rate of plants in each condition}
\usage{
GetGrowthRate()
}
\value{
Data frame
}
\description{
Calculate the growth rate of plants in each condition
}
