% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GrowthRateData}
\alias{GrowthRateData}
\title{Growth rate calculation from RawWeightData}
\format{
A dataframe with 10 variables and 72 obervations
\describe{
  \item{genotype}{Genotype}
  \item{photoperiod}{Photoperiod}
  \item{intensity}{Light intensity (μmol per second per squared meter)}
  \item{age}{Plant age (day)}
  \item{mean_weight}{Mean weight (g)}
  \item{instant_growth_rate}{Instantaneous growth rate (g per day)}
  \item{avg_growth_rate}{Average growth rate since germination (g per day)}
  \item{photon_per_day}{Photon intensity per day (mol per day per squared meter)}
  \item{instant_growth_rate_per_photon}{Instantaneous growth per light intensity (g m2 per mol)}
  \item{averageg_growth_rate_per_photon}{Average growth per light intensity (g m2 per mol)}
}
}
\usage{
GrowthRateData
}
\description{
Growth rate calculation from RawWeightData
}
\keyword{datasets}
