% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhotonEfficiencyPlot.R
\name{PhotonEfficiencyPlot}
\alias{PhotonEfficiencyPlot}
\title{Plot a parallel coordinated plot of the photon efficiency data}
\usage{
PhotonEfficiencyPlot()
}
\value{
ggplot2 object
}
\description{
Plot a parallel coordinated plot of the photon efficiency data
}
