% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGrowthRate.R
\name{PlotGrowthRate}
\alias{PlotGrowthRate}
\title{Plot a faceted plot of the growth rate data}
\usage{
PlotGrowthRate(
  instant_or_average = "instant",
  facet_scale = "fixed",
  log_rate = FALSE,
  perPhoton = FALSE
)
}
\arguments{
\item{instant_or_average}{"instant" or "average". Whether instantaneous or average growth rate should be plotted.}

\item{facet_scale}{"free", "free_x", "free_y" or "fixed".}

\item{log_rate}{Logical. If TRUE then the growth rate is log10 transformed.}

\item{perPhoton}{Logical. If TRUE then the per photon growth rate is plotted.}
}
\value{
ggplot2 object
}
\description{
Plot a faceted plot of the growth rate data
}
