% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRawWeight.R
\name{PlotRawWeight}
\alias{PlotRawWeight}
\title{Plot a faceted plot of the raw weight data}
\usage{
PlotRawWeight(facet_scale = "fixed", log_rate = FALSE)
}
\arguments{
\item{facet_scale}{"free", "free_x", "free_y" or "fixed".}

\item{log_rate}{Logical. If TRUE then the growth rate is log10 transformed.}
}
\value{
ggplot2 object
}
\description{
Plot a faceted plot of the raw weight data
}
