% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RatePlot.R
\name{RatePlot}
\alias{RatePlot}
\title{Plot a faceted plot of the growth rate data}
\usage{
RatePlot(log_rate = FALSE, select_genotype = "WT", value = "weight")
}
\arguments{
\item{log_rate}{Logical. If TRUE then the growth rate is log10 transformed.}

\item{select_genotype}{Character. One of "WT", "co9", "ft10", "mips1".}

\item{value}{Character. One of "weight", "average_rate", "instant_rate".}
}
\value{
ggplot2 object
}
\description{
Plot a faceted plot of the growth rate data
}
