% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RatePlotPanel.R
\name{RatePlotPanel}
\alias{RatePlotPanel}
\title{Plot all three types of rate plots at once.}
\usage{
RatePlotPanel(select_genotype = "WT", log_rate = FALSE)
}
\arguments{
\item{select_genotype}{Character. One of "WT", "co9", "ft10", "mips1".}

\item{log_rate}{Logical. If TRUE then the growth rate is log10 transformed.}
}
\value{
ggplot2 object
}
\description{
Plot all three types of rate plots at once.
}
