% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RawWeightData}
\alias{RawWeightData}
\title{Raw Data of an Experiment Aimed to Determine Plant Growth Rate}
\format{
A data frame with 360 observations and 5 variables:
\describe{
  \item{genotype}{Genotype}
  \item{photoperiod}{Photoperiod}
  \item{intensity}{Light intensity (μmol photon per second per squared meter)}
  \item{age}{Plant age (day)}
  \item{weight}{Weight (g)}
}
}
\usage{
RawWeightData
}
\description{
Raw Data of an Experiment Aimed to Determine Plant Growth Rate
}
\keyword{datasets}
